Refactor: auth type 수정
