+++ src/types/auth/model.type.ts
@@ -1,0 +1,12 @@
+export type KakaoPropertiesModel = {
+  nickname: string;
+  profile_image: string;
+  thumbnail_image: string;
+};
+
+export type KakaoAccountModel = {
+  email: string;
+  age_range: string;
+  birthday: string;
+  gender: string;
+};
