+++ src/types/auth/response.type.ts
@@ -1,24 +1,13 @@
-export type AuthResponseModel = {
+import { KakaoAccountModel, KakaoPropertiesModel } from '@/types/auth/model.type';
+
+export type AuthResponse = {
   accessToken: string;
   refreshToken: string;
   userId: number;
   accessTokenExpireDate: number;
 };
 
-export type KakaoPropertiesModel = {
-  nickname: string;
-  profile_image: string;
-  thumbnail_image: string;
-};
-
-export type KakaoAccountModel = {
-  email: string;
-  age_range: string;
-  birthday: string;
-  gender: string;
-};
-
-export type KakaoUserInfoResponseModel = {
+export type KakaoUserInfoResponse = {
   id: string;
   properties: KakaoPropertiesModel;
   kakao_account: KakaoAccountModel;
